requirements_test.txt
# Dependencies
django-model-utils>=2.0
monero>=0.6
python-monerorpc>=0.5.5
djangorestframework>=3.9.3

# Test Dependencies
coverage==4.4.1
mock>=1.0.1
flake8>=2.1.0
tox>=1.7.0
coveralls
pytest
pytest-django
pytest-cov
model_mommy
